docs: 补充 README 项目截图与模拟数据
- 新增 seed_data.sql，包含 15 篇文章、13 条评论
- 修复 init.sql 外键检查，支持重复初始化
- README 新增项目展示章节与 5 张页面截图
- 同步当前前后端功能迭代

+++ novablog-server/src/main/java/com/novablog/entity/Article.java
@@ -39,12 +39,12 @@ public class Article {
     /**
      * 浏览量
      */
-    private Integer viewCount;
+    private Long viewCount;
 
     /**
      * 点赞数
      */
-    private Integer likeCount;
+    private Long likeCount;
 
     /**
      * 作者ID
